refactor: rename extension example files for consistency
