+++ doc/src/js/examples/extend-animation.js
