+++ doc/src/js/examples/extend-fly-controls.js
