+++ doc/src/js/examples/extend-procedural-data-provider.js
