+++ doc/src/js/examples/extend-three-renderer.js
