+++ doc/src/js/examples/extend-webgl-renderer.js
