+++ doc/src/pages/examples/extend-procedural-data-provider.js
@@ -11,10 +11,10 @@ import React from 'react';
 import Layout from '@theme/Layout';
 import {Code} from '../../js/components/Code';
 import {ViewerComponent} from '../../js/components/ViewerComponent';
-import {dispose, init} from '../../js/examples/three-renderer';
+import {dispose, init} from '../../js/examples/extend-procedural-data-provider';
 
 export default function Example() {
-  const title = 'Three Renderer';
+  const title = 'Extend Procedural Data Provider';
   return (
     <Layout title={title} noFooter="true">
       <ViewerComponent init={init} dispose={dispose} />
