+++ doc/src/pages/examples/extend-webgl-renderer.js
@@ -11,10 +11,10 @@ import React from 'react';
 import Layout from '@theme/Layout';
 import {Code} from '../../js/components/Code';
 import {ViewerComponent} from '../../js/components/ViewerComponent';
-import {dispose, init} from '../../js/examples/animation';
+import {dispose, init} from '../../js/examples/extend-webgl-renderer';
 
 export default function Example() {
-  const title = 'Animation';
+  const title = 'Extend WebGL Renderer';
   return (
     <Layout title={title} noFooter="true">
       <ViewerComponent init={init} dispose={dispose} />
