+++ examples/doc/extend-animation.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
     <head>
-        <title>Fly Controls</title>
+        <title>Extend Animation</title>
         <link rel="icon" href="data:," />
         <meta charset="utf-8" />
         <meta
@@ -29,7 +29,7 @@
     <body>
         <script type="module">
             import { accessToken } from "/doc-src/.access-token/token.js";
-            import { init } from "/doc-src/src/js/examples/fly-controls.js";
+            import { init } from "/doc-src/src/js/examples/extend-animation.js";
 
             const container = document.createElement("div");
             container.className = "viewer";
