+++ examples/doc/extend-fly-controls.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
     <head>
-        <title>WebGL Renderer</title>
+        <title>Extend Fly Controls</title>
         <link rel="icon" href="data:," />
         <meta charset="utf-8" />
         <meta
@@ -29,7 +29,7 @@
     <body>
         <script type="module">
             import { accessToken } from "/doc-src/.access-token/token.js";
-            import { init } from "/doc-src/src/js/examples/webgl-renderer.js";
+            import { init } from "/doc-src/src/js/examples/extend-fly-controls.js";
 
             const container = document.createElement("div");
             container.className = "viewer";
