+++ examples/doc/extend-webgl-renderer.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
     <head>
-        <title>Animation</title>
+        <title>Extend WebGL Renderer</title>
         <link rel="icon" href="data:," />
         <meta charset="utf-8" />
         <meta
@@ -29,7 +29,7 @@
     <body>
         <script type="module">
             import { accessToken } from "/doc-src/.access-token/token.js";
-            import { init } from "/doc-src/src/js/examples/animation.js";
+            import { init } from "/doc-src/src/js/examples/extend-webgl-renderer.js";
 
             const container = document.createElement("div");
             container.className = "viewer";
